Feature update

+++ backend/main.py
@@ -9,7 +9,9 @@ if backend_dir not in sys.path:
     sys.path.insert(0, backend_dir)
     
 from utils.point_cloud_utils import visualize_point_cloud
+from utils.file_operations import find_processed_file
 from point_cloud_processor import steps_to_process_point_cloud
+
 
 # Function to process the point cloud
 def process(original_path, destination_directory):
@@ -17,22 +19,29 @@ def process(original_path, destination_directory):
         os.makedirs(destination_directory)
 
     filename = os.path.basename(original_path)
-    destination_path = os.path.join(destination_directory, filename)
+    
+    existing_file = find_processed_file(filename, destination_directory)
+    if existing_file is None:
+        destination_path = os.path.join(destination_directory, filename)
 
-    shutil.copyfile(original_path, destination_path)
+        shutil.copyfile(original_path, destination_path)
+    else:
+        #A existing file is in the destination, so update to use that one. 
+        destination_path = existing_file
 
     processed_point_cloud = steps_to_process_point_cloud(destination_path, destination_directory)
 
     return processed_point_cloud
 
 # Function to visualize the point cloud
-def visualize(path_to_visualize):
-    visualize_point_cloud(path_to_visualize)
+def visualize(path_to_visualize, original_path = None):
+    visualize_point_cloud(path_to_visualize, original_path)
 
 # Function to process and then visualize the point cloud
 def process_and_visualize(original_path, destination_directory):
+    starting_point_cloud = original_path
     processed_file_path = process(original_path, destination_directory)
-    visualize(processed_file_path)
+    visualize(processed_file_path, starting_point_cloud)
 
 def main():
     parser = argparse.ArgumentParser(description="Process and/or visualize point cloud files.")
